reorganize storm package
