+++ summingbird-storm/src/main/scala/com/twitter/summingbird/storm/Constants.scala
@@ -14,11 +14,10 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package com.twitter.summingbird
+package com.twitter.summingbird.storm
 
-import com.twitter.summingbird.option._
+import com.twitter.summingbird.option.MonoidIsCommutative
 import com.twitter.summingbird.storm.option._
-import com.twitter.summingbird.util.CacheSize
 
 object Constants {
   val AGG_KEY     = "aggKey"
