+++ summingbird-storm/src/main/scala/com/twitter/summingbird/storm/FutureQueue.scala
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package com.twitter.summingbird.util
+package com.twitter.summingbird.storm
 
 import com.twitter.util.{ Await, Future }
 import scala.collection.mutable.{ Queue => MutableQueue }
