+++ summingbird-storm/src/main/scala/com/twitter/summingbird/util/package.scala
@@ -1,0 +1,25 @@
+/*
+ Copyright 2013 Twitter, Inc.
+
+ Licensed under the Apache License, Version 2.0 (the "License");
+ you may not use this file except in compliance with the License.
+ You may obtain a copy of the License at
+
+ http://www.apache.org/licenses/LICENSE-2.0
+
+ Unless required by applicable law or agreed to in writing, software
+ distributed under the License is distributed on an "AS IS" BASIS,
+ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ See the License for the specific language governing permissions and
+ limitations under the License.
+ */
+
+package com.twitter.summingbird
+
+package object util {
+  @deprecated("Use com.twitter.summingbird.storm.option.CacheSize", "0.1.0")
+  type CacheSize = storm.option.CacheSize
+
+  @deprecated("Use com.twitter.summingbird.storm.option.CacheSize", "0.1.0")
+  val CacheSize = storm.option.CacheSize
+}
